// src/components/select-sort/enum.ts
enum TagNames {
  DIV = 'div',
  LABEL = 'label',
  SELECT = 'select',
  OPTION = 'option',
}

enum Styles {
  CONTAINER = 'select-container',
  LABEL = 'label',
  SELECT = 'select-sort',
  OPTION = 'option',
  SELECT_ERROR = 'select--error',
}

enum Events {
  CHANGE = 'change',
}

enum Address {
  BILLING = 'billing',
  SHIPPING = 'shipping',
}

enum Attributes {
  ID = 'id',
  FOR = 'for',
  NAME = 'name',
  VALUE = 'value',
}

export { TagNames, Styles, Address, Attributes, Events };